+++ spec/GeoJsonDumper.spec.ts
@@ -3,13 +3,15 @@ import GeoJsonDumper, { GeoJson } from "GeoJsonDumper";
 
 describe("GeoJSON Dumper", () => {
   const geocoded = Geocoded.create({
-    latitude: 38.8978378,
-    longitude: -77.0365123,
+    coordinates: {
+      latitude: 38.8978378,
+      longitude: -77.0365123,
+    },
     bounds: {
-      latitude1: 38.89380528242933,
-      longitude1: -77.04317326462667,
-      latitude2: 38.90153071757068,
-      longitude2: -77.02993873537334,
+      latitudeSW: 38.89380528242933,
+      longitudeSW: -77.04317326462667,
+      latitudeNE: 38.90153071757068,
+      longitudeNE: -77.02993873537334,
     },
     formattedAddress:
       "1600 Pennsylvania Avenue Northwest, Washington, DC 20050",
@@ -25,12 +27,12 @@ describe("GeoJSON Dumper", () => {
       type: "Point",
       coordinates: [-77.0365123, 38.8978378],
     },
-    bounds: {
-      latitude1: 38.89380528242933,
-      longitude1: -77.04317326462667,
-      latitude2: 38.90153071757068,
-      longitude2: -77.02993873537334,
-    },
+    bbox: [
+      -77.04317326462667,
+      38.89380528242933,
+      -77.02993873537334,
+      38.90153071757068,
+    ],
   };
 
   it("provides a valid GeoJSON dump", () => {