+++ spec/provider/BingProvider.spec.ts
@@ -54,12 +54,15 @@ describe("Bing Geocoder Provider", () => {
         const geocoded = results[0];
 
         expect(geocoded).toBeDefined();
-        expect(geocoded.getCoordinates()).toEqual([38.897668, -77.036556]);
+        expect(geocoded.getCoordinates()).toEqual({
+          latitude: 38.897668,
+          longitude: -77.036556,
+        });
         expect(geocoded.getBounds()).toEqual({
-          latitude1: 38.89380528242933,
-          longitude1: -77.04317326462667,
-          latitude2: 38.90153071757068,
-          longitude2: -77.02993873537334,
+          latitudeSW: 38.89380528242933,
+          longitudeSW: -77.04317326462667,
+          latitudeNE: 38.90153071757068,
+          longitudeNE: -77.02993873537334,
         });
         expect(geocoded.getFormattedAddress()).toEqual(
           "1600 Pennsylvania Ave NW, Washington, DC 20006"
@@ -90,12 +93,15 @@ describe("Bing Geocoder Provider", () => {
       const geocoded = results[0];
 
       expect(geocoded).toBeDefined();
-      expect(geocoded.getCoordinates()).toEqual([48.8631093, 2.3887809]);
+      expect(geocoded.getCoordinates()).toEqual({
+        latitude: 48.8631093,
+        longitude: 2.3887809,
+      });
       expect(geocoded.getBounds()).toEqual({
-        latitude1: 48.85924658242932,
-        longitude1: 2.380952653445271,
-        latitude2: 48.866972017570674,
-        longitude2: 2.396609146554729,
+        latitudeSW: 48.85924658242932,
+        longitudeSW: 2.380952653445271,
+        latitudeNE: 48.866972017570674,
+        longitudeNE: 2.396609146554729,
       });
       expect(geocoded.getFormattedAddress()).toEqual(
         "8 Avenue Gambetta, 75020 Paris"