+++ spec/provider/MapboxProvider.spec.ts
@@ -52,7 +52,10 @@ describe("Mapbox Geocoder Provider", () => {
       const geocoded = results[0];
 
       expect(geocoded).toBeDefined();
-      expect(geocoded.getCoordinates()).toEqual([38.87925, -76.98204]);
+      expect(geocoded.getCoordinates()).toEqual({
+        latitude: 38.87925,
+        longitude: -76.98204,
+      });
       expect(geocoded.getBounds()).toEqual(undefined);
       expect(geocoded.getFormattedAddress()).toEqual(
         "1600 Pennsylvania Ave SE, Washington, District of Columbia 20003, United States"
@@ -90,7 +93,10 @@ describe("Mapbox Geocoder Provider", () => {
       const geocoded = results[0];
 
       expect(geocoded).toBeDefined();
-      expect(geocoded.getCoordinates()).toEqual([48.863134, 2.388886]);
+      expect(geocoded.getCoordinates()).toEqual({
+        latitude: 48.863134,
+        longitude: 2.388886,
+      });
       expect(geocoded.getBounds()).toEqual(undefined);
       expect(geocoded.getFormattedAddress()).toEqual(
         "12 Avenue Gambetta, 75020 Paris, France"