+++ spec/provider/NominatimProvider.spec.ts
@@ -54,12 +54,15 @@ describe("OpenStreetMap / Nominatim Geocoder Provider", () => {
         const geocoded = results[0];
 
         expect(geocoded).toBeDefined();
-        expect(geocoded.getCoordinates()).toEqual([38.8958536, -77.0307129]);
+        expect(geocoded.getCoordinates()).toEqual({
+          latitude: 38.8958536,
+          longitude: -77.0307129,
+        });
         expect(geocoded.getBounds()).toEqual({
-          latitude1: 38.8957842,
-          longitude1: -77.0309688,
-          latitude2: 38.895924,
-          longitude2: -77.0304609,
+          latitudeSW: 38.8957842,
+          longitudeSW: -77.0309688,
+          latitudeNE: 38.895924,
+          longitudeNE: -77.0304609,
         });
         expect(geocoded.getFormattedAddress()).toEqual(undefined);
         expect(geocoded.getDisplayName()).toEqual(
@@ -104,15 +107,15 @@ describe("OpenStreetMap / Nominatim Geocoder Provider", () => {
         const geocoded = results[0];
 
         expect(geocoded).toBeDefined();
-        expect(geocoded.getCoordinates()).toEqual([
-          48.863744499999996,
-          2.3911562136123106,
-        ]);
+        expect(geocoded.getCoordinates()).toEqual({
+          latitude: 48.863744499999996,
+          longitude: 2.3911562136123106,
+        });
         expect(geocoded.getBounds()).toEqual({
-          latitude1: 48.8625929,
-          longitude1: 2.3877078,
-          latitude2: 48.8648832,
-          longitude2: 2.3956964,
+          latitudeSW: 48.8625929,
+          longitudeSW: 2.3877078,
+          latitudeNE: 48.8648832,
+          longitudeNE: 2.3956964,
         });
         expect(geocoded.getFormattedAddress()).toEqual(undefined);
         expect(geocoded.getDisplayName()).toEqual(