+++ spec/provider/OpenCageProvider.spec.ts
@@ -52,12 +52,15 @@ describe("OpenCage Geocoder Provider", () => {
       const geocoded = results[0];
 
       expect(geocoded).toBeDefined();
-      expect(geocoded.getCoordinates()).toEqual([38.8636383, -76.9463651]);
+      expect(geocoded.getCoordinates()).toEqual({
+        latitude: 38.8636383,
+        longitude: -76.9463651,
+      });
       expect(geocoded.getBounds()).toEqual({
-        latitude1: 38.8633822,
-        longitude1: -76.9467576,
-        latitude2: 38.8637409,
-        longitude2: -76.945632,
+        latitudeSW: 38.8633822,
+        longitudeSW: -76.9467576,
+        latitudeNE: 38.8637409,
+        longitudeNE: -76.945632,
       });
       expect(geocoded.getFormattedAddress()).toEqual(
         "Pennsylvania Avenue, Washington, DC 20746-8001, United States of America"
@@ -101,7 +104,10 @@ describe("OpenCage Geocoder Provider", () => {
       const geocoded = results[0];
 
       expect(geocoded).toBeDefined();
-      expect(geocoded.getCoordinates()).toEqual([48.863116, 2.38878]);
+      expect(geocoded.getCoordinates()).toEqual({
+        latitude: 48.863116,
+        longitude: 2.38878,
+      });
       expect(geocoded.getBounds()).toEqual(undefined);
       expect(geocoded.getFormattedAddress()).toEqual(
         "8 Avenue Gambetta, 75020 Paris, France"