+++ src/provider/googlemaps/GoogleMapsProvider.ts
@@ -233,10 +233,10 @@ export default class GoogleMapsProvider
       {
         address: geocodeQuery.getText(),
         bounds: geocodeQuery.getBounds()
-          ? `${geocodeQuery.getBounds()?.latitude1},${
-              geocodeQuery.getBounds()?.longitude1
-            }|${geocodeQuery.getBounds()?.latitude2},${
-              geocodeQuery.getBounds()?.longitude2
+          ? `${geocodeQuery.getBounds()?.latitudeSW},${
+              geocodeQuery.getBounds()?.longitudeSW
+            }|${geocodeQuery.getBounds()?.latitudeNE},${
+              geocodeQuery.getBounds()?.longitudeNE
             }`
           : undefined,
         components: (<GoogleMapsGeocodeQuery>geocodeQuery).getComponents()
@@ -552,8 +552,10 @@ export default class GoogleMapsProvider
     });
 
     let geocoded = GoogleMapsGeocoded.create({
-      latitude,
-      longitude,
+      coordinates: {
+        latitude,
+        longitude,
+      },
       formattedAddress,
       streetNumber,
       streetName,
@@ -588,26 +590,26 @@ export default class GoogleMapsProvider
     if (result.geometry.bounds) {
       const { bounds } = result.geometry;
       geocoded = <GoogleMapsGeocoded>geocoded.withBounds({
-        latitude1: bounds.southwest.lat,
-        longitude1: bounds.southwest.lng,
-        latitude2: bounds.northeast.lat,
-        longitude2: bounds.northeast.lng,
+        latitudeSW: bounds.southwest.lat,
+        longitudeSW: bounds.southwest.lng,
+        latitudeNE: bounds.northeast.lat,
+        longitudeNE: bounds.northeast.lng,
       });
     } else if (result.geometry.viewport) {
       const { viewport } = result.geometry;
       geocoded = <GoogleMapsGeocoded>geocoded.withBounds({
-        latitude1: viewport.southwest.lat,
-        longitude1: viewport.southwest.lng,
-        latitude2: viewport.northeast.lat,
-        longitude2: viewport.northeast.lng,
+        latitudeSW: viewport.southwest.lat,
+        longitudeSW: viewport.southwest.lng,
+        latitudeNE: viewport.northeast.lat,
+        longitudeNE: viewport.northeast.lng,
       });
     } else if (result.geometry.location_type === "ROOFTOP") {
       // Fake bounds
       geocoded = <GoogleMapsGeocoded>geocoded.withBounds({
-        latitude1: latitude,
-        longitude1: longitude,
-        latitude2: latitude,
-        longitude2: longitude,
+        latitudeSW: latitude,
+        longitudeSW: longitude,
+        latitudeNE: latitude,
+        longitudeNE: longitude,
       });
     }